+++ cmd/get/clusters/runner_test.go
@@ -42,10 +42,8 @@ func Test_run(t *testing.T) {
 		{
 			name: "case 0: get clusters",
 			storage: []runtime.Object{
-				newcapiCluster("1sad2", "10.5.0", "some-org", "test cluster 3", label.ServicePriorityHighest, parseCreated("default"), nil),
-				newAWSClusterResource("1sad2", "10.5.0", "some-org", "test cluster 3", creationTime, nil),
-				newcapiCluster("f930q", "11.0.0", "some-other", "test cluster 4", label.ServicePriorityMedium, parseCreated("default"), nil),
-				newAWSClusterResource("f930q", "11.0.0", "some-other", "test cluster 4", creationTime, nil),
+				newcapiCluster("1sad2", "10.5.0", "some-org", "test cluster 3", label.ServicePriorityHighest, creationTime, nil),
+				newcapiCluster("f930q", "11.0.0", "some-other", "test cluster 4", label.ServicePriorityMedium, creationTime, nil),
 			},
 			args:               nil,
 			expectedGoldenFile: "run_get_clusters.golden",
@@ -60,9 +58,7 @@ func Test_run(t *testing.T) {
 			name: "case 2: get cluster by id",
 			storage: []runtime.Object{
 				newcapiCluster("1sad2", "10.5.0", "some-org", "test cluster 3", label.ServicePriorityHighest, creationTime, nil),
-				newAWSClusterResource("1sad2", "10.5.0", "some-org", "test cluster 3", creationTime, nil),
 				newcapiCluster("f930q", "11.0.0", "some-other", "test cluster 4", label.ServicePriorityMedium, creationTime, nil),
-				newAWSClusterResource("f930q", "11.0.0", "some-other", "test cluster 4", creationTime, nil),
 			},
 			args:               []string{"f930q"},
 			expectedGoldenFile: "run_get_cluster_by_id.golden",
@@ -70,16 +66,6 @@ func Test_run(t *testing.T) {
 		{
 			name:         "case 3: get cluster by id, with empty storage",
 			storage:      nil,
-			args:         []string{"f930q"},
-			errorMatcher: IsNotFound,
-		},
-		{
-			name: "case 4: get cluster by id, with no infrastructure cluster",
-			storage: []runtime.Object{
-				newcapiCluster("1sad2", "10.5.0", "some-org", "test cluster 3", label.ServicePriorityHighest, parseCreated("default"), nil),
-				newAWSClusterResource("1sad2", "10.5.0", "some-org", "test cluster 3", parseCreated("2021-01-01T15:04:32Z"), nil),
-				newcapiCluster("f930q", "11.0.0", "some-other", "test cluster 3", label.ServicePriorityMedium, parseCreated("default"), nil),
-			},
 			args:         []string{"f930q"},
 			errorMatcher: IsNotFound,
 		},