+++ cmd/get/nodepools/runner_test.go
@@ -7,11 +7,14 @@ import (
 	"time"
 
 	"github.com/giantswarm/k8sclient/v8/pkg/k8sclienttest"
+	"github.com/giantswarm/k8smetadata/pkg/label"
 	"github.com/giantswarm/microerror"
 	"github.com/google/go-cmp/cmp"
 	"github.com/pkg/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
+	capi "sigs.k8s.io/cluster-api/api/core/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake" //nolint:staticcheck
 
 	"github.com/giantswarm/kubectl-gs/v5/internal/key"
@@ -43,9 +46,7 @@ func Test_run(t *testing.T) {
 			name: "case 0: get nodepools",
 			storage: []runtime.Object{
 				newcapiMachineDeployment("1sad2", "s921a", "10.5.0", creationTime, 2, 1),
-				newAWSMachineDeployment("1sad2", "s921a", "10.5.0", "test nodepool 3", creationTime, 1, 3),
 				newcapiMachineDeployment("f930q", "s921a", "11.0.0", creationTime, 6, 6),
-				newAWSMachineDeployment("f930q", "s921a", "11.0.0", "test nodepool 4", creationTime, 5, 8),
 			},
 			args:               nil,
 			expectedGoldenFile: "run_get_nodepools.golden",
@@ -59,9 +60,7 @@ func Test_run(t *testing.T) {
 			name: "case 2: get nodepool by name",
 			storage: []runtime.Object{
 				newcapiMachineDeployment("1sad2", "s921a", "10.5.0", creationTime, 2, 1),
-				newAWSMachineDeployment("1sad2", "s921a", "10.5.0", "test nodepool 3", creationTime, 1, 3),
 				newcapiMachineDeployment("f930q", "s921a", "11.0.0", creationTime, 6, 6),
-				newAWSMachineDeployment("f930q", "s921a", "11.0.0", "test nodepool 4", creationTime, 5, 8),
 			},
 			args:               []string{"f930q"},
 			expectedGoldenFile: "run_get_nodepool_by_id.golden",
@@ -72,51 +71,35 @@ func Test_run(t *testing.T) {
 			errorMatcher: IsNotFound,
 		},
 		{
-			name: "case 4: get nodepool by name, with no infrastructure ref",
+			name: "case 4: get nodepools by cluster name",
 			storage: []runtime.Object{
 				newcapiMachineDeployment("1sad2", "s921a", "10.5.0", creationTime, 2, 1),
-				newAWSMachineDeployment("1sad2", "s921a", "10.5.0", "test nodepool 3", creationTime, 1, 3),
 				newcapiMachineDeployment("f930q", "s921a", "11.0.0", creationTime, 6, 6),
-			},
-			args:         []string{"f930q"},
-			errorMatcher: IsNotFound,
-		},
-		{
-			name: "case 5: get nodepools by cluster name",
-			storage: []runtime.Object{
-				newcapiMachineDeployment("1sad2", "s921a", "10.5.0", creationTime, 2, 1),
-				newAWSMachineDeployment("1sad2", "s921a", "10.5.0", "test nodepool 3", creationTime, 1, 3),
-				newcapiMachineDeployment("f930q", "s921a", "11.0.0", creationTime, 6, 6),
-				newAWSMachineDeployment("f930q", "s921a", "11.0.0", "test nodepool 4", creationTime, 5, 8),
 				newcapiMachineDeployment("9f012", "29sa0", "9.0.0", creationTime, 0, 3),
-				newAWSMachineDeployment("9f012", "29sa0", "9.0.0", "test nodepool 5", creationTime, 1, 1),
 			},
 			args:               nil,
 			clusterName:        "s921a",
 			expectedGoldenFile: "run_get_nodepool_by_cluster_id.golden",
 		},
 		{
-			name: "case 6: get nodepools by name and cluster name",
+			name: "case 5: get nodepools by name and cluster name",
 			storage: []runtime.Object{
 				newcapiMachineDeployment("1sad2", "s921a", "10.5.0", creationTime, 2, 1),
-				newAWSMachineDeployment("1sad2", "s921a", "10.5.0", "test nodepool 3", creationTime, 1, 3),
 				newcapiMachineDeployment("f930q", "s921a", "11.0.0", creationTime, 6, 6),
-				newAWSMachineDeployment("f930q", "s921a", "11.0.0", "test nodepool 4", creationTime, 5, 8),
 				newcapiMachineDeployment("9f012", "29sa0", "9.0.0", creationTime, 0, 3),
-				newAWSMachineDeployment("9f012", "29sa0", "9.0.0", "test nodepool 5", creationTime, 1, 1),
 			},
 			args:               []string{"f930q"},
 			clusterName:        "s921a",
 			expectedGoldenFile: "run_get_nodepool_by_id_and_cluster_id.golden",
 		},
 		{
-			name:               "case 7: get nodepools by cluster name, with empty storage",
+			name:               "case 6: get nodepools by cluster name, with empty storage",
 			args:               nil,
 			clusterName:        "s921a",
 			expectedGoldenFile: "run_get_nodepool_by_cluster_id_empty_storage.golden",
 		},
 		{
-			name:         "case 8: get nodepools by name and cluster name, with empty storage",
+			name:         "case 7: get nodepools by name and cluster name, with empty storage",
 			args:         []string{"f930q"},
 			clusterName:  "s921a",
 			errorMatcher: IsNotFound,
@@ -139,7 +122,7 @@ func Test_run(t *testing.T) {
 				service:      newClusterService(t, tc.storage...),
 				flag:         flag,
 				stdout:       out,
-				provider:     key.ProviderAWS,
+				provider:     key.ProviderDefault,
 			}
 
 			err := runner.run(ctx, nil, tc.args)
@@ -176,6 +159,33 @@ func Test_run(t *testing.T) {
 			}
 		})
 	}
+}
+
+func newcapiMachineDeployment(name, clusterName, release string, creationDate time.Time, nodesDesired, nodesReady int) *capi.MachineDeployment {
+	n := &capi.MachineDeployment{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "cluster.x-k8s.io/v1beta1",
+			Kind:       "MachineDeployment",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:              name,
+			Namespace:         "default",
+			CreationTimestamp: metav1.NewTime(creationDate),
+			Labels: map[string]string{
+				label.MachineDeployment: name,
+				label.ReleaseVersion:    release,
+				label.Organization:      "giantswarm",
+				label.Cluster:           clusterName,
+				capi.ClusterNameLabel:   clusterName,
+			},
+		},
+		Status: capi.MachineDeploymentStatus{
+			Replicas:      int32(nodesDesired), //nolint:gosec
+			ReadyReplicas: int32(nodesReady),   //nolint:gosec
+		},
+	}
+
+	return n
 }
 
 func newClusterService(t *testing.T, object ...runtime.Object) *nodepool.Service {